requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.11
# by the following command:
#
#    pip-compile --output-file=requirements.txt pyproject.toml
#
click==8.1.7
    # via leapfrogai
confz==2.0.1
    # via leapfrogai
diskcache==5.6.3
    # via llama-cpp-python
grpcio==1.59.3
    # via
    #   grpcio-health-checking
    #   grpcio-reflection
    #   leapfrogai
grpcio-health-checking==1.59.3
    # via leapfrogai
grpcio-reflection==1.59.3
    # via leapfrogai
leapfrogai==0.4.0rc1
    # via leapfrogai-backend-llama-cpp-py (pyproject.toml)
llama-cpp-python==0.2.20
    # via leapfrogai-backend-llama-cpp-py (pyproject.toml)
numpy==1.26.2
    # via llama-cpp-python
protobuf==4.25.1
    # via
    #   grpcio-health-checking
    #   grpcio-reflection
    #   leapfrogai
pydantic==1.10.13
    # via
    #   confz
    #   leapfrogai
python-dotenv==1.0.0
    # via confz
pyyaml==6.0.1
    # via confz
toml==0.10.2
    # via confz
typing-extensions==4.8.0
    # via
    #   llama-cpp-python
    #   pydantic
